Combined phase index lists works correctly
In stead of outputing two sets of file, one for negative and one for
positive phases the anicr now outputs one set of file where the phase
has been encoded in the third index as an extra sign. This reduces the
number of files on disk most of which are empty anyway.